Move dialogs and windows to their own modules
Re #182
