+++ mantidimaging/core/filters/roi_normalisation/roi_normalisation_gui.py
@@ -1,6 +1,6 @@
 from functools import partial
 
-from mantidimaging.gui.stack_visualiser import Parameters
+from mantidimaging.gui.windows.stack_visualiser import Parameters
 
 from . import execute
