+++ mantidimaging/gui/dialogs/async_task/model.py
