+++ mantidimaging/gui/dialogs/async_task/presenter.py
