+++ mantidimaging/gui/dialogs/async_task/view.py
