+++ mantidimaging/gui/dialogs/filters/__init__.py
