+++ mantidimaging/gui/dialogs/filters/navigation_toolbar.py
