+++ mantidimaging/gui/dialogs/filters/presenter.py
