+++ mantidimaging/gui/dialogs/filters/view.py
