+++ mantidimaging/gui/utility/__init__.py
@@ -5,6 +5,8 @@ from .get_auto_params_from_stack import (  # noqa: F401
 
 from .qt_helpers import (  # noqa: F401
         compile_ui,
+        select_file,
+        select_directory,
         add_property_to_form,
         delete_all_widgets_from_layout,
         BlockQtSignals
