+++ mantidimaging/gui/utility/qt_helpers.py
@@ -40,6 +40,35 @@ def compile_ui(ui_file, qt_obj=None):
     return uic.loadUi(os.path.join(base_path, ui_file), qt_obj)
 
 
+def select_file(field, caption):
+    """
+    :param field: The field in which the result will be saved
+    :param caption: Title of the file browser window that will be opened
+    :return: True: If a file has been selected, False otherwise
+    """
+    assert isinstance(field, Qt.QLineEdit), (
+            "The passed object is of type {0}. This function only works with "
+            "QLineEdit".format(type(field)))
+
+    selected_file = Qt.QFileDialog.getOpenFileName(caption=caption)[0]
+    # open file dialogue and set the text if file is selected
+    if selected_file:
+        field.setText(selected_file)
+        return True
+
+    # no file has been selected
+    return False
+
+
+def select_directory(field, caption):
+    assert isinstance(field, Qt.QLineEdit), (
+            "The passed object is of type {0}. This function only works with "
+            "QLineEdit".format(type(field)))
+
+    # open file dialogue and set the text if file is selected
+    field.setText(Qt.QFileDialog.getExistingDirectory(caption=caption))
+
+
 def add_property_to_form(label,
                          dtype,
                          default_value=None,
@@ -82,7 +111,6 @@ def add_property_to_form(label,
 
     # Set up data type dependant widgets
     if dtype == 'file':
-        from mantidimaging.gui.main_window.load_dialog import select_file
         left_widget = Qt.QLineEdit()
         right_widget = Qt.QPushButton(label)
         assign_tooltip([left_widget, right_widget])
