+++ mantidimaging/gui/windows/main/model.py
