+++ mantidimaging/gui/windows/main/test/model_test.py
@@ -2,7 +2,7 @@ import unittest
 
 from mantidimaging.core.utility.special_imports import import_mock
 
-from mantidimaging.gui.main_window import MainWindowModel
+from mantidimaging.gui.windows.main import MainWindowModel
 
 mock = import_mock()
