+++ mantidimaging/gui/windows/main/test/presenter_test.py
@@ -2,9 +2,9 @@ import unittest
 
 from mantidimaging.core.utility.special_imports import import_mock
 
-from mantidimaging.gui.async_task_dialog import TaskWorkerThread
-from mantidimaging.gui.main_window.load_dialog import MWLoadDialog
-from mantidimaging.gui.main_window import (
+from mantidimaging.gui.dialogs.async_task import TaskWorkerThread
+from mantidimaging.gui.windows.main.load_dialog import MWLoadDialog
+from mantidimaging.gui.windows.main import (
         MainWindowView, MainWindowPresenter)
 
 mock = import_mock()
