+++ mantidimaging/gui/windows/stack_visualiser/__init__.py
