+++ mantidimaging/gui/windows/stack_visualiser/histogram.py
