+++ mantidimaging/gui/windows/stack_visualiser/navigation_toolbar.py
