+++ mantidimaging/gui/windows/stack_visualiser/presenter.py
