+++ mantidimaging/gui/windows/stack_visualiser/roi_selector_widget.py
