+++ mantidimaging/gui/windows/stack_visualiser/test/__init__.py
