+++ mantidimaging/gui/windows/stack_visualiser/view.py
